FIShBOTR: The Title of the Project

This packages provides R helper functions to make use of the FIShBOT dataset. Fishing Industry Shared Bottom Oceanographic Timeseries (FIShBOT) is a standardized in-situ gridded bottom temperature product derived from cooperative research data streams. The data are collected by commercial fishermen operating in US waters of the Northwest Atlantic Ocean. To improve efficiency, the data are stored as a table online. This package converts the data back to a grid and provides a number of functions to facilitate analysis.

See also

Author

Maintainer: George Maynard george.maynard@noaa.gov (ORCID)

Authors:

Sarah Salois sarah.salois@noaa.gov